# frontend/client/openapi.json
{"openapi":"3.1.0","info":{"title":"Quark Investimentos API","version":"0.0.1"},"paths":{"/test":{"get":{"summary":"Test","operationId":"test_test_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth":{"post":{"tags":["Authorization"],"summary":"Login","operationId":"login_auth_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tokens"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Authorization"],"summary":"Refresh Token","operationId":"refresh_token_auth_patch","parameters":[{"name":"refresh_token","in":"query","required":true,"schema":{"type":"string","title":"Refresh Token"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tokens"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Authorization"],"summary":"Logout","operationId":"logout_auth_delete","responses":{"204":{"description":"Successful Response"}}}},"/users/{role_id}":{"post":{"tags":["User Management"],"summary":"Register User","operationId":"register_user_users__role_id__post","parameters":[{"name":"role_id","in":"path","required":true,"schema":{"type":"integer","title":"Role Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserForm"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"integer","title":"Response Register User Users  Role Id  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/operations/import":{"post":{"tags":["Operation Management"],"summary":"Import Operations","operationId":"import_operations_operations_import_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_operations_operations_import_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Operation Management"],"summary":"Approve Operations","operationId":"approve_operations_operations_import_patch","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/operations/report/active":{"get":{"tags":["Operation Management"],"summary":"Generate Active Operations Report","operationId":"generate_active_operations_report_operations_report_active_get","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Generate Active Operations Report Operations Report Active Get"}}}}}}},"/operations/report/current":{"get":{"tags":["Operation Management"],"summary":"Generate Current Operations Report","operationId":"generate_current_operations_report_operations_report_current_get","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Generate Current Operations Report Operations Report Current Get"}}}}}}},"/clients/import":{"post":{"tags":["Clients Management"],"summary":"Import Clients","operationId":"import_clients_clients_import_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_clients_clients_import_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Clients Management"],"summary":"Approve Clients","operationId":"approve_clients_clients_import_patch","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/clients/report":{"get":{"tags":["Clients Management"],"summary":"Generate My Clients Report","operationId":"generate_my_clients_report_clients_report_get","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Generate My Clients Report Clients Report Get"}}}}}}}},"components":{"schemas":{"Body_import_clients_clients_import_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_import_clients_clients_import_post"},"Body_import_operations_operations_import_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_import_operations_operations_import_post"},"Credentials":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"Credentials"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Tokens":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"expires_in":{"type":"integer","title":"Expires In"},"token_type":{"type":"string","title":"Token Type"},"timestamp":{"type":"number","title":"Timestamp"}},"type":"object","required":["access_token","refresh_token","expires_in","token_type","timestamp"],"title":"Tokens"},"UserForm":{"properties":{"credentials":{"$ref":"#/components/schemas/Credentials"},"name":{"type":"string","title":"Name"}},"type":"object","required":["credentials","name"],"title":"UserForm"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"Authorization","description":"Grant Access."},{"name":"Authentication","description":"Verify Identity."},{"name":"User","description":"User Operations"}]}